+++ backend/canisters/community/api/src/updates/c2c_join_channel.rs
@@ -2,7 +2,7 @@ use candid::{CandidType, Principal};
 use serde::{Deserialize, Serialize};
 use types::{
     ChannelId, CommunityCanisterChannelSummary, CommunityCanisterCommunitySummary, GateCheckFailedReason, TimestampMillis,
-    UniquePersonProof, UserId, VerifiedCredentialGateArgs,
+    UniquePersonProof, UserId, UserType, VerifiedCredentialGateArgs,
 };
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
@@ -13,6 +13,8 @@ pub struct Args {
     pub invite_code: Option<u64>,
     pub is_platform_moderator: bool,
     pub is_bot: bool,
+    #[serde(default)]
+    pub user_type: UserType,
     pub diamond_membership_expires_at: Option<TimestampMillis>,
     pub verified_credential_args: Option<VerifiedCredentialGateArgs>,
     pub unique_person_proof: Option<UniquePersonProof>,