+++ backend/canisters/user/impl/src/model/direct_chat.rs
@@ -5,9 +5,10 @@ use serde::{Deserialize, Serialize};
 use std::cmp::min;
 use types::{
     DirectChatSummary, DirectChatSummaryUpdates, EventWrapper, Message, MessageId, MessageIndex, Milliseconds, OptionUpdate,
-    TimestampMillis, Timestamped, UserId,
+    TimestampMillis, Timestamped, UserId, UserType,
 };
 use user_canister::SendMessageArgs;
+use utils::consts::OPENCHAT_BOT_USER_ID;
 
 #[derive(Serialize, Deserialize)]
 pub struct DirectChat {
@@ -20,13 +21,21 @@ pub struct DirectChat {
     pub notifications_muted: Timestamped<bool>,
     pub archived: Timestamped<bool>,
     pub is_bot: bool,
+    #[serde(default)]
+    pub user_type: UserType,
     pub unconfirmed: Vec<SendMessageArgs>,
 }
 
 impl DirectChat {
+    pub fn set_user_type(&mut self) {
+        if self.is_bot {
+            self.user_type = if self.them == OPENCHAT_BOT_USER_ID { UserType::OcControlledBot } else { UserType::Bot };
+        }
+    }
+
     pub fn new(
         them: UserId,
-        is_bot: bool,
+        user_type: UserType,
         events_ttl: Option<Milliseconds>,
         anonymized_chat_id: u128,
         now: TimestampMillis,
@@ -40,7 +49,8 @@ impl DirectChat {
             read_by_them_up_to: Timestamped::new(None, now),
             notifications_muted: Timestamped::new(false, now),
             archived: Timestamped::new(false, now),
-            is_bot,
+            is_bot: user_type.is_bot(),
+            user_type,
             unconfirmed: Vec::new(),
         }
     }