+++ backend/canisters/user_index/impl/src/lifecycle/post_upgrade.rs
@@ -1,11 +1,13 @@
 use crate::lifecycle::{init_env, init_state};
 use crate::memory::get_upgrades_memory;
-use crate::Data;
+use crate::{mutate_state, Data};
 use canister_logger::LogEntry;
 use canister_tracing_macros::trace;
 use ic_cdk::post_upgrade;
+use local_user_index_canister::UserRegistered;
 use stable_memory::get_reader;
 use tracing::info;
+use types::UserType;
 use user_index_canister::post_upgrade::Args;
 use utils::cycles::init_cycles_dispenser_client;
 
@@ -22,6 +24,20 @@ fn post_upgrade(args: Args) {
     let env = init_env(data.rng_seed, data.oc_key_pair.is_initialised());
     init_cycles_dispenser_client(data.cycles_dispenser_canister_id, data.test_mode);
     init_state(env, data, args.wasm_version);
+
+    mutate_state(|state| {
+        state.push_event_to_all_local_user_indexes(
+            local_user_index_canister::Event::UserRegistered(UserRegistered {
+                user_id: state.data.proposals_bot_canister_id.into(),
+                user_principal: state.data.proposals_bot_canister_id,
+                username: "ProposalsBot".to_string(),
+                is_bot: true,
+                user_type: UserType::OcControlledBot,
+                referred_by: None,
+            }),
+            None,
+        );
+    });
 
     info!(version = %args.wasm_version, "Post-upgrade complete");
 }