+++ backend/canisters/user_index/impl/src/model/user_map.rs
@@ -7,7 +7,7 @@ use serde::{Deserialize, Serialize};
 use std::collections::{BTreeSet, HashMap};
 use std::ops::RangeFrom;
 use tracing::info;
-use types::{CyclesTopUp, Milliseconds, SuspensionDuration, TimestampMillis, UniquePersonProof, UserId};
+use types::{CyclesTopUp, Milliseconds, SuspensionDuration, TimestampMillis, UniquePersonProof, UserId, UserType};
 use utils::case_insensitive_hash_map::CaseInsensitiveHashMap;
 use utils::time::MonthKey;
 
@@ -61,12 +61,12 @@ impl UserMap {
         username: String,
         now: TimestampMillis,
         referred_by: Option<UserId>,
-        is_bot: bool,
+        user_type: UserType,
     ) {
         self.username_to_user_id.insert(&username, user_id);
         self.principal_to_user_id.insert(principal, user_id);
 
-        let user = User::new(principal, user_id, username, now, referred_by, is_bot);
+        let user = User::new(principal, user_id, username, now, referred_by, user_type);
         self.users.insert(user_id, user);
 
         if let Some(ref_by) = referred_by {
@@ -379,7 +379,7 @@ impl UserMap {
             user.username.clone(),
             user.date_created,
             None,
-            false,
+            UserType::User,
         );
         self.update(user, date_created, false);
     }
@@ -451,9 +451,9 @@ mod tests {
         let user_id2: UserId = Principal::from_slice(&[3, 2]).into();
         let user_id3: UserId = Principal::from_slice(&[3, 3]).into();
 
-        user_map.register(principal1, user_id1, username1.clone(), 1, None, false);
-        user_map.register(principal2, user_id2, username2.clone(), 2, None, false);
-        user_map.register(principal3, user_id3, username3.clone(), 3, None, false);
+        user_map.register(principal1, user_id1, username1.clone(), 1, None, UserType::User);
+        user_map.register(principal2, user_id2, username2.clone(), 2, None, UserType::User);
+        user_map.register(principal3, user_id3, username3.clone(), 3, None, UserType::User);
 
         let principal_to_user_id: Vec<_> = user_map
             .principal_to_user_id
@@ -490,7 +490,7 @@ mod tests {
 
         let user_id = Principal::from_slice(&[1, 1]).into();
 
-        user_map.register(principal, user_id, username1, 1, None, false);
+        user_map.register(principal, user_id, username1, 1, None, UserType::User);
 
         if let Some(original) = user_map.get_by_principal(&principal) {
             let mut updated = original.clone();