+++ backend/canisters/user_index/impl/src/updates/c2c_register_bot.rs
@@ -3,7 +3,7 @@ use canister_tracing_macros::trace;
 use event_store_producer::EventBuilder;
 use ic_cdk::update;
 use local_user_index_canister::{Event, UserRegistered};
-use types::{Cycles, UserId};
+use types::{Cycles, UserId, UserType};
 use user_index_canister::c2c_register_bot::{Response::*, *};
 use utils::text_validation::{validate_username, UsernameValidationError};
 
@@ -48,7 +48,7 @@ fn c2c_register_bot_impl(args: Args, state: &mut RuntimeState) -> Response {
     state
         .data
         .users
-        .register(caller, user_id, args.username.clone(), now, None, true);
+        .register(caller, user_id, args.username.clone(), now, None, UserType::Bot);
 
     state.push_event_to_all_local_user_indexes(
         Event::UserRegistered(UserRegistered {
@@ -56,6 +56,7 @@ fn c2c_register_bot_impl(args: Args, state: &mut RuntimeState) -> Response {
             user_principal: caller,
             username: args.username,
             is_bot: true,
+            user_type: UserType::Bot,
             referred_by: None,
         }),
         None,